refactor(structure): phase 3 — fold UI buckets
- entities/* (Testimonials, RelatedPosts, BlogPostsGrid) → components/
- features/* (ThemeSelector, LivePreviewListener, PayloadRedirects, LocaleSelector) → components/
- widgets compositions colocate: BlogPageContent → blog/_components, PostContent → blog/[slug]/_components
- Header/Footer now import from their collections directly (dissolved widgets barrel)
- providers/BeforeOpenDrawerWrapper → lib/context
- src root now: app, blocks, collections, components, globals, lib, messages
