+++ apps/cms/src/app/(frontend)/[locale]/blog/[slug]/_components/PostContent/components/RelatedPostsSection.tsx
