+++ apps/cms/src/app/(frontend)/[locale]/blog/_components/BlogPageContent/components/FeaturedPost.tsx
