+++ apps/cms/src/app/(frontend)/[locale]/blog/_components/BlogPageContent/types.ts
