--- apps/cms/src/entities/index.ts
@@ -1,3 +1,0 @@
-export { BlogPostsGrid } from "./BlogPostsGrid";
-export { RelatedPosts } from "./RelatedPosts";
-export { AnimatedCarousel, TestimonialCard } from "./Testimonials";
