--- apps/cms/src/features/index.ts
@@ -1,3 +1,0 @@
-export { ThemeSelector } from "./ThemeSelector";
-export { LivePreviewListener } from "./LivePreviewListener";
-export { PayloadRedirects } from "./PayloadRedirects";
